Re-organized Elasticsearch and DatabaseCleaner rspec configuration
This configuration has been moved to separate files in an effort to begin cleaning spec_helper.rb.
